move playthrough issue-related objects to improvements
