+++ core/templates/dev/head/services/PlaythroughIssuesBackendApiService.ts
@@ -15,8 +15,8 @@
 /**
  * @fileoverview Service for fetching issues and playthroughs from the backend.
  */
+require('domain/improvements/PlaythroughIssueObjectFactory.ts');
 require('domain/statistics/PlaythroughObjectFactory.ts');
-require('domain/statistics/PlaythroughIssueObjectFactory.ts');
 require('domain/utilities/UrlInterpolationService.ts');
 
 require('services/services.constants.ts');
